+++ src/meowmx/__init__.py
@@ -1,13 +1,25 @@
 from .client import Client, ExpectedVersionFailure
-from .common import Engine, NewEvent, NewEventRow, RecordedEvent, Session, SessionMaker
+from .common import (
+    Engine,
+    EventCompatible,
+    NewEvent,
+    NewEventRow,
+    RecordedEvent,
+    Session,
+    SessionMaker,
+)
+from .aggregates import EventBuffer, PendingEvents
 
 __all__ = [
     "Client",
     "Engine",
+    "EventBuffer",
+    "EventCompatible",
     "ExpectedVersionFailure",
     "NewEvent",
     "NewEventRow",
     "RecordedEvent",
+    "PendingEvents",
     "Session",
     "SessionMaker",
 ]